feat(updated): added util for daos(update)
