+++ server/daos/authorsBooks.js
@@ -1,14 +1,14 @@
 import db from '@database/models';
 
-export const insertAuthorsBooks = async args => db.authorsBooks.bulkCreate(args);
+export const insertAuthorsBooks = args => db.authorsBooks.bulkCreate(args);
 
-export const updateAuthorsBooksForBooks = async args =>
+export const updateAuthorsBooksForBooks = args =>
   db.authorsBooks.bulkCreate(args, {
     fields: ['id', 'bookId', 'authorId'],
     updateOnDuplicate: ['bookId']
   });
 
-export const updateAuthorsBooksForAuthors = async args =>
+export const updateAuthorsBooksForAuthors = args =>
   db.authorsBooks.bulkCreate(args, {
     fields: ['id', 'bookId', 'authorId'],
     updateOnDuplicate: ['authorId']
